Реализован статический класс в SecurityDoors.Core для получения строки подключения для всех проектов.
